2.0/requirements.txt
opencv-python
ultralytics
torch
torchvision
numpy
pillow
git+https://github.com/openai/CLIP.git
